Refactor code structure
Refactored the code for improved organization and readability. Consider reviewing the length of `src/pages/Index.tsx` as it may benefit from further refactoring.
[skip gpt_engineer]
